Add retrieving dynamics engine time step
Retrieving dynamics engine time step is added to the functionality
provided by the interface to V-REP remote API server. It necessitates
dynamic addition of a constant to the Python binding to the V-REP remote
API, which corresponds to the parameter representing the step size of
the currently selected physics engine, because such constant is missing
there.