Add retrieving whether V-REP simulation is started
Retrieving whether V-REP simulation is started is added to the
functionality provided by the interface to V-REP remote API server. It
may give inaccurate results if it is invoked immediately after starting
or stopping a simulation; in such a case, introducing a short delay
before invoking it should help. Because the V-REP remote API does not
provide the relevant functionality itself, retrieving whether V-REP
simulation is started is based on inspecting the information on the
server state bundled with the last message received from the remote API
server (to ensure that this information is up-to-date, prior to
inspection a new message is fetched from the remote API server by
retrieving whether V-REP is currently waiting for a trigger signal).