Add setting object position
Setting object position is added to the functionality provided by the
interface to a generic scene object simulated in V-REP. By default, to
avoid decomposition of dynamically simulated objects, setting object
position is not allowed during simulation (but this restriction can be
overridden). It also necessitates addition of an attribute storing the
interface to V-REP remote API server.