Change optional argument for setting/retrieving scene object orientation
An optional argument for retrieving scene object orientation, which
indicates a relative reference frame, is changed such that it can also
be specified as an instance of class SceneObject or subclass thereof. A
similar change is made to an optional argument for setting scene object
orientation.

Additionally, because this functionality requires resolving the type of
the argument and since similar type resolving is necessary for one of
the arguments used for setting scene object parent, the common code is
moved to a separate static function.