Change connecting to and disconnecting from V-REP remote API server
Connecting to and disconnecting from V-REP remote API server is changed
such that establishing multiple connections is explicitly forbidden.
It necessitates addition of a global (module-level) variable that keeps
track of whether a connection has already been established (because only
one instance of class Simulator can maintain a connection at any given
time, but many instances of this class can be created).

Additionally, because an error encountered when connecting to V-REP
remote API server is considered a different condition from that when the
return code of an invoked remote API function indicates an error, one
exception is added:

- ConnectionError - raised during connecting to V-REP remote API server.

It is adopted wherever appropriate.