Change hierarchy of classes that communicate with V-REP simulator
The hierarchy of classes that communicate with V-REP simulator is
changed such that classes SceneObject, Collection, and NengoComm become
subclasses of class Communicator. Consequently, all references to the
appropriate interface to V-REP remote API server as well as to the
client ID of this interface are obtained in these classes and their
subclasses using respective properties of class Communicator.