Change handling errors when calling V-REP remote API functions
Handling errors when calling V-REP remote API functions is changed such
that two conditions are distinguished:

- when a function cannot be called because a connection to V-REP remote
  API server is missing, in which case exception ConnectionError is
  raised;
- when a function is called but it fails, in which case exception
  ServerError is raised.

It necessitates an appropriate change to the semantics and consequently
to the docstring of exception ConnectionError. Additionally, the
docstrings of other exceptions are also changed to make them consistent
with the docstring of exception ConnectionError.