Change argument for interfaces communicating with V-REP simulator
The argument for each interface communicating with V-REP simulator,
which specifies an interface to V-REP remote API server, is changed to
optional. If it is not specified, a default interface to V-REP remote
API server is obtained via parent class Communicator. It also
necessitates changing the position of the argument on the argument list,
and thus the last position is adopted.