Change arguments for interface to V-REP remote API server
Arguments for interface to V-REP remote API server are changed such that
arguments specifying the following information are added:

- blocking wait until establishing connection with V-REP remote API
  server or exceeding timeout;
- automatic attempts to reconnect to V-REP remote API server after
  losing connection;
- timeout for establishing connection with V-REP remote API server or
  for blocking function calls;
- interval between data exchanges with V-REP.

Consequently, specifying the above information when connecting to V-REP
remote API server is no longer needed and therefore the relevant
arguments in this case are removed.

Additionally, the semantics of the argument specifying automatic
attempts to reconnect to V-REP remote API server after losing connection
is changed from negative to positive meaning, which is also reflected by
an appropriate change in the name of the argument.