Add finalizer to interface to V-REP remote API server
A finalizer is added to the interface to V-REP remote API server. It
disconnects the interface from the V-REP remote API server if the
interface is connected. However, it does not properly reset other
variables, so deleting an instance of an interface to V-REP remote API
server should not be used for closing a connection. The finalizer is
considered only as a protective measure for the event when the interface
to V-REP remote API server is still connected while quitting Python.