Change handling scene objects/models removed from scene
Handling scene objects that have been removed from scene is changed such
that exception RuntimeError is raised whenever any property of an
instance of an interface to such a scene object is accessed or any
method of an instance of such an interface is invoked. Similar changes
are made to handling models that have been removed from scene.

Additionally, handling the name of a removed scene object/model is
changed such that the name is maintained upon removal so that it could
be used in error messages. Consequently, it necessitates providing other
means for distinguishing whether a scene object/model has been removed.
For this purpose, the internal representation of the object handle is
adopted, which is changed to assume a distinct value in such cases.