Change arguments for interfaces to scene objects and models in V-REP
Arguments for interfaces to scene objects simulated in V-REP, to models
simulated in V-REP, and to arrays of scene objects simulated in V-REP
are changed such that a new argument representing an interface to the
parent scene object (or parent model) is added. The position on the
argument list adopted for this argument is last but one. It also
necessitates addition of an attribute storing the interface to the
parent scene object/model.

Additionally, setting object parent is adapted to appropriately update
the attribute storing the interface to the parent scene object/model as
well as registered child objects of the old and new parents.