Add retrieving distance from proximity sensor
Retrieving distance to the detected point is added to the functionality
provided by the interface to proximity sensor simulated in V-REP.
Distance can be retrieved either in a fast manner, which involves
directly returning solely the z coordinate of the detected point, or in
an exact manner, in which case the Euclidean distance to the detected
point is computed based on all of its 3 coordinates and then returned.
In either case, the retrieved distance is expressed in the relative
reference frame of the proximity sensor.