Change rounding float values returned by V-REP remote API functions
Rounding float values returned by several functions of the V-REP remote
API is changed such that by default no rounding is performed. Instead,
rounding may be carried out on demand, but in this case it is necessary
to explicitly specify the desired precision. Consequently, the
specification of the default precision for all relevant functions
scattered across different modules is no longer needed and, as a result,
the global constant specifying such precision is removed from the module
storing global constants.